Added check for account_suspensions.conf Include
I've added a new function - check_for_local_template_account_suspensions_include - which checks for account_suspensions.conf if the user has a main.local Apache template. If the suspensions include is missing, it will provide a warning that a main.local exists and that it is missing the account_suspensions.conf include.