Convert ssp script to modulino for easier function testing
This converts the script to a modulino.  Which means, you
can now write a new function, then execute just that function
without having to run the entire application.

For a good explanation of what a modulino is, please read:
 http://perltricks.com/article/107/2014/8/7/Rescue-legacy-code-with-modulinos

In doing this, you can now write unit tests.  In addition,
it encoruages new new developers to use better coding practices.
It also helps to ensure that when globals exist, it will show
how important it is to comment their use.  Finally, capitalizing
the globals is an important step into helping newer developers
implement a better coding practice of distinguishing between
local variables, from global ones.

Finally, this commit helps me so that I can implement a PHP
FPM checker, and be able to test just my script, without worry
that I broke something else.

NOTE: You'll notice that there's code changes where just the
format of the code was changed.  This is due to me running
cPanel's perltidy.