Remove expensive, useless rpm check
Case NO-1166: On EasyApache4 systems, the questions of "what version
of EasyApache" and "when was Apache last built" are not relevant.
Instead of running a potentially expensive RPM query to generate
meaningless data to fill space in a template, just leave it out.

On a system with platter hard drives, where the RPM database is not
already cached in RAM, this saves about 10 seconds.