Don't repeat ourselves when using abs_path getting symlink abs_path
The prior (removed) ternary would always find true due to the minimum
variable being '/' rather than undef.  As such, it is sensible to
simply abs_path the computed full path directly.

Furthermore, there is no reason to allocate anything extra for the
computed full path given:

1. the path will always be able to be split into an array w/ nonzero
length save in the case of /; and if / is a symlink on the system
It is highly unlikely that SSP will be able to gather useful data
in any capacity.

2. Referring to an array slice as done here will never warn, even
on 0-length arrays.  Nor will join() or split().

3. Given the above ternary always returning true, we should simply
set the variable.